// Front End/API-faturamento/src/components/LoginForm/types.ts
export interface FormProps {
  width?: string;
}

export interface StyledButtonGroupProps {
  position?: string;
  display?: string;
  alignitems?: string;
  justifycontent?: string;
  margintop?: string;
}

export interface StyledInputGroupProps {
  position?: string;
  display?: string;
  alignitems?: string;
}

export interface PasswordIconToggleProps {
  position?: string;
  right?: string;
  border?: string;
  background?: string;
  cursor?: string;
}